--- AppleMusicInspired/LibraryScreen/RadioScreen/RadioStationCell.swift
@@ -1,18 +1,0 @@
-//
-//  RadioStationCell.swift
-//  AppleMusicInspired
-//
-//  Created by Oksana Kazarinova on 20/09/2024.
-//
-
-import SwiftUI
-
-struct RadioStationCell: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    RadioStationCell()
-}
