+++ AppleMusicInspired/RadioScreen/CollectionDataModel.swift
